+++ examples/save_load_dock_state.rs
@@ -23,6 +23,10 @@ struct TabViewer {
 impl egui_dock::TabViewer for TabViewer {
     type Tab = String;
 
+    fn id(&mut self, tab: &mut Self::Tab) -> egui::Id {
+        egui::Id::new(tab)
+    }
+
     fn title(&mut self, tab: &mut Self::Tab) -> egui::WidgetText {
         (&*tab).into()
     }